### 2018-08-16 #896 @whedon
[PRE REVIEW]: hoggorm: a python library for explorative multivariate statistics
**Submitting author:** @olivertomic (<a href="http://orcid.org/0000-0003-1595-9962">Oliver Tomic</a>)
**Repository:** <a href="https://github.com/olivertomic/hoggorm" target ="_blank">https://github.com/olivertomic/hoggorm</a>
**Version:** 0.12.0
