### 2018-08-16 #896 @arfon
Hi @olivertomic, please remove the `orcid` YAML key for authors who don't have one (e.g. Thomas Graff)
