### 2018-08-16 e36896c
Removed empty orcid id's
