+++ paper.md
@@ -12,13 +12,11 @@ authors:
    orcid: 0000-0003-1595-9962
    affiliation: 1
  - name: Thomas Graff
-   orcid: none
    affiliation: 2
  - name: Kristian Hovde Liland
    orcid: 0000-0001-6468-9423
    affiliation: 1
  - name: Tormod Næs
-   orcid: none
    affiliation: 3
 affiliations:
 - name: Norwegian University of Life Sciences
@@ -27,7 +25,7 @@ affiliations:
   index: 2
 - name: Nofima
   index: 3
-date: 09 July 2018
+date: 16 August 2018
 bibliography: paper.bib
 ---
 
