### 2018-08-16 #896 @olivertomic
Hi @arfon, thank you very much for initialising the process and the feedback regarding the orcid YAML keys. I removed them for two of the authors that were not registered with orcid.
