### 2018-09-17 #896 @arfon
@olivertomic - could you take a look at this [list of potential reviewers](https://bit.ly/joss-reviewers) and suggest some people who might be good candidates?
