### 2018-09-17 #896 @olivertomic
@arfon, thank you for getting in touch.  A very good candidate would be candidate on row 338 on your list, since this person is the only one who has named chemometrics as their domain. Another great candidate would be candidate on row 275. Other possible candidates would be 108, 123, 173.
