### 2018-09-17 #896 @arfon
@expectopatronum @GalenStocking @tupi @jsgalan would any of you be willing to review this submission for JOSS? You can learn more about what that process looks like here: https://joss.readthedocs.io/en/latest/review_criteria.html
