### 2018-09-26 #896 @arfon
@jsgalan @olivertomic - please head over to #980 for the actual review
