### 2018-10-27 #980 @jsgalan
Hi all,

I am sorry for my belated  response.  Here are my initial review comments:

-Please make sure to explain the acronyms, are TGXnet and Nofima Companies?
-In the reference from Martens and Naes, please check the title "Principal component regression in NIR analysis: Viewpoints, background details and selection of components", and add the DOI accordingly, check [site](https://onlinelibrary.wiley.com/doi/abs/10.1002/cem.1180020207)

Best
