### 2018-10-30 3764a9c
Updating references for JOSS
- Inserted missing doi's if available
- Fixed article reference naes88 (updated order of authors and title) 
- Inserted missing ISBN
