+++ paper.bib
@@ -4,6 +4,7 @@
   publisher = {London: Academic Press},
   year = {1979},
   author = {K.V. Mardia and J.T. Kent and J.M. Bibby}
+  isbn = {9780124712508}
 }
 
 @INBOOK{wold82,
@@ -11,17 +12,19 @@
   title = {Systems under Indirect Observation},
   publisher = {Amsterdam: North Holland},
   year = {1982},
-  editor = {K.G. Jøreskog and H. Wold},
-  author = {H. Wold}
+  editor = {K.G. J{\"o}reskog and H. Wold},
+  author = {H. Wold},
+  isbn = {9780444863010}
 }
 
-@ARTICLE{martens88,
-  author = {Harald Martens and Tormod N{\ae}s},
-  title = {Principal components regression in NIR analysis},
+@ARTICLE{naes88,
+  author = {Tormod N{\ae}s, and Harald Martens},
+  title = {Principal components regression in {NIR} analysis: Viewpoints, background details and selection of components},
   journal = {Journal of Chemometrics},
   year = {1988},
   volume = {2},
-  pages = {155-167}
+  pages = {155-167},
+  doi = {10.1002/cem.1180020207 }
 }
 
 @ARTICLE{robert76,
@@ -30,21 +33,23 @@
   journal = {Applied Statistics},
   year = {1976},
   volume = {25},
-  pages = {257–265}
+  pages = {257–265}, 
+  doi = {10.2307/2347233}
 }
 
 
 @ARTICLE{smilde09,
-  author = {Smilde, Age K. and Kiers, Henk A. L. and Bijlsma, S. and Rubingh, C. M. and van Erk, M. J.},
+  author = {A.K. Smilde and H.A.L. Kiers  and S. Bijlsma and C.M. Rubingh and M.J. van Erk},
   title = {Matrix correlations for high-dimensional data: the modified RV-coefficient},
   journal = {Bioinformatics},
   year = {2009},
   volume = {25},
   pages = {401-405},
+  doi = {10.1093/bioinformatics/btn634}
 }
 
 @ARTICLE{indahl18,
-  author = {U.G. Indahl, T. N{\ae}s, K.H. Liland},
+  author = {U.G. Indahl and T. N{\ae}s and K.H. Liland},
   title = {A similarity index for comparing coupled matrices},
   journal = {Journal of Chemometrics},
   year = {2018},
