### 2018-10-30 b8f6cd9
Updates for JOSS journal
Updated reference for article "Principal component regression in NIR analysis: Viewpoints, background details and selection of components".
