### 2018-10-30 #980 @olivertomic
@jsgalan thank you very much your first comments and for taking on the reviewer job.

1. I have updated the reference. Thank you for bringing my attention to this. I also inserted DOI's and ISBN where those were missing.
2. TGXnet is a (coding) company in Norway and Nofima is a food research institute in Norway. Where and how would you like me to insert this information? In paper.md in the affiliation part?

I will make @whedon produce a new pdf with the updated references. I will ask @whedon to generate another pdf once I got your feedback regarding point 2.

best
