+++ paper.bib
@@ -3,7 +3,7 @@
   title = {Multivariate Analysis},
   publisher = {London: Academic Press},
   year = {1979},
-  author = {K.V. Mardia and J.T. Kent and J.M. Bibby}
+  author = {K.V. Mardia and J.T. Kent and J.M. Bibby},
   isbn = {9780124712508}
 }
 
@@ -24,7 +24,7 @@
   year = {1988},
   volume = {2},
   pages = {155-167},
-  doi = {10.1002/cem.1180020207 }
+  doi = {10.1002/cem.1180020207}
 }
 
 @ARTICLE{robert76,
