### 2018-11-10 #980 @jsgalan
Hi all,

I overall think is a great package that will simplify a few routinely task done (by us) chemometricians.

_**ISSUES ABOUT THE SOFTWARE PACKAGE**_

i have successfully installed both packages **hoggorm** and **hoggormplot**.
![screen shot 2018-11-10 at 2 37 29 pm](https://user-images.githubusercontent.com/971958/48305528-22b45280-e4fa-11e8-9b73-a748153641ec.png)
which is good, but still I have a few suggestions to be made:

- add the examples to the package itself, both as .py and .ipynb
- separate each example type via a folder and put each examples under a suitable and explanatory filename. For instance in the Partial Least Squares Regression example there are 3 examples:

1. Partial Least Squares regression in sensometrics
2. Partial Least Squares Regression in spectroscopy
3. Partial Least Squares regression with air pollution data

The notebooks are great for teaching purposes in the webpage, but it still need to be in code with a proper documentation.

- I see there is a 'tests'  foldre in the package, please explain the usage of this examples in the main documentation

- Please Specify in the documentation the matplotlib and numpy version used to compile the package. For instance, I am using Matplotlib 2.2.0 and numy 1.15.0, and have issues with closing the figures for no apparent reason.

- In the issue of plots, could you add an example where 2 plots are in the same figure(), i am not sure if i should use subplot from matplotlib or a specific command in your package.

- I will be happy to see a PLS2 example

_**ISSUES ABOUT THE PAPER AND SUBMISSION**_
- I am unsure the license is following an OSI license type (https://opensource.org/licenses/alphabetical)

- In the article itself i think it should be addressed more clearly the issues that this softwares solves. It is evident to me as a chemometrician familiar with the methods, but in my experience not many fields relate to the use of PSLR as their methods of data summarization and explanation.

- I see you guys are using NIPALS algorithm please add the appropriate citation for the algorithm (as reference please check: http://www.vias.org/tmdatanaleng/dd_nipals_algo.html)
- Add more information about how and when to see special  matrix correlation coefficients RV, RV2 and SMI.
- There is a typo in the main hoggorm page ( https://github.com/olivertomic/hoggorm) it reads matrix **corrlation** coefficients RV, RV2 and SMI where it should be "correlation".

Best
