### 2018-11-12 #980 @olivertomic
Hi @jsgalan! Thanks a lot for your effort and providing feedback on the package. I will try to implement all requirements as quickly as possible and get back to you once I am done.

best wishes
