### 2019-01-18 #980 @jsgalan
Hi @olivertomic, I will be happy to check those changes !

Best
