### 2019-02-15 28d035f
New examples for PCA
Included two new Jupyter notebooks with examples of how to apply PCA on data.
