+++ examples/data/cheese_sensory.txt
@@ -1,0 +1,14 @@
+6.19	3.33	3.43	2.14	1.29	3.11	6.7	3.22	2.66	5.1	4.57	3.34	2.93	1.89	1.23	3.15	4.07
+6.55	2.5	4.32	2.52	1.24	3.91	6.68	2.57	2.42	4.87	4.75	4.13	3.09	2.29	1.51	3.93	4.07
+6.23	3.43	3.42	2.03	1.28	2.93	6.61	3.39	2.56	5	4.73	3.44	3.08	1.81	1.37	3.19	4.16
+6.14	2.93	3.96	2.13	1.08	3.12	6.51	2.98	2.5	4.66	4.68	3.92	2.93	1.99	1.19	3.13	4.29
+6.7	1.97	4.72	2.43	1.13	4.6	7.01	2.07	2.32	5.29	5.19	4.52	3.14	2.47	1.34	4.67	4.03
+6.19	5.28	1.59	1.07	1	1.13	6.42	5.18	2.82	5.02	4.49	2.05	2.54	1.18	1.18	1.29	4.11
+6.17	3.45	3.32	2.04	1.47	2.69	6.39	3.81	2.76	4.58	4.32	3.22	2.72	1.81	1.33	2.52	4.26
+6.9	2.58	4.24	2.58	1.7	4.19	7.11	2.06	2.47	4.58	5.09	4.44	3.25	2.62	1.73	4.87	3.98
+6.7	2.53	4.53	2.32	1.22	4.16	6.91	2.42	2.41	4.52	4.96	4.49	3.37	2.47	1.64	4.54	4.01
+6.35	3.14	3.64	2.17	1.17	2.57	6.5	2.77	2.66	4.76	4.64	4.06	3.11	2.21	1.46	3.35	3.93
+5.97	3.34	3.46	1.67	1.15	1.43	6.31	3.15	2.56	4.57	4.36	3.65	2.66	1.56	1.19	2.23	4.01
+6.29	2.99	4.03	2.06	1.17	3.06	6.76	2.37	2.44	4.69	4.97	4.28	3.16	2.56	1.53	4.23	4.03
+5.91	4.88	2.04	1	1	1.08	6.34	4.79	2.44	5.48	4.54	1.98	2.57	1	1.03	1.03	4.16
+6.75	1.91	4.36	2.95	1.43	4.83	7.14	1.53	2.47	4.72	5.06	4.54	3.43	2.8	1.87	5.65	3.98
