### 2019-02-16 e26f6af
Added examples for PCR and PLSR
