+++ examples/PCR_on_sensory_and_fluorescence_data.py
@@ -1,0 +1,563 @@
+
+# coding: utf-8
+
+# # Principal Component Regression (PCR) on Sensory and Fluorescence data
+
+# This notebook illustrates how to use the **hoggorm** package to carry out principal component regression (PCR) on multivariate data. Furthermore, we will learn how to visualise the results of the PCR using the **hoggormPlot** package.
+
+# ---
+
+# ### Import packages and prepare data
+
+# First import **hoggorm** for analysis of the data and **hoggormPlot** for plotting of the analysis results. We'll also import **pandas** such that we can read the data into a data frame. **numpy** is needed for checking dimensions of the data.
+
+# In[1]:
+
+
+import hoggorm as ho
+import hoggormplot as hop
+import pandas as pd
+import numpy as np
+
+
+# Next, load the data that we are going to analyse using **hoggorm**. After the data has been loaded into the pandas data frame, we'll display it in the notebook.
+
+# In[2]:
+
+
+# Load fluorescence data
+X_df = pd.read_csv('cheese_fluorescence.txt', index_col=0, sep='\t')
+X_df
+
+
+# In[3]:
+
+
+# Load sensory data
+Y_df = pd.read_csv('cheese_sensory.txt', index_col=0, sep='\t')
+Y_df
+
+
+# The ``nipalsPCR`` class in hoggorm accepts only **numpy** arrays with numerical values and not pandas data frames. Therefore, the pandas data frames holding the imported data need to be "taken apart" into three parts: 
+# * two numpy array holding the numeric values
+# * two Python list holding variable (column) names
+# * two Python list holding object (row) names. 
+# 
+# The numpy arrays with values will be used as input for the ``nipalsPCR`` class for analysis. The Python lists holding the variable and row names will be used later in the plotting function from the **hoggormPlot** package when visualising the results of the analysis. Below is the code needed to access both data, variable names and object names.
+
+# In[4]:
+
+
+# Get the values from the data frame
+X = X_df.values
+Y = Y_df.values
+
+# Get the variable or columns names
+X_varNames = list(X_df.columns)
+Y_varNames = list(Y_df.columns)
+
+# Get the object or row names
+X_objNames = list(X_df.index)
+Y_objNames = list(Y_df.index)
+
+
+# ---
+
+# ### Apply PCR to our data
+
+# Now, let's run PCR on the data using the ``nipalsPCR`` class. The documentation provides a [description of the input parameters](https://hoggorm.readthedocs.io/en/latest/pcr.html). Using input paramter ``arrX`` and ``arrY`` we define which numpy array we would like to analyse. ``arrY`` is what typically is considered to be the response matrix, while the measurements are typically defined as ``arrX``. By setting input parameter ``Xstand=False`` and ``Ystand=False`` we make sure that the variables are only mean centered, not scaled to unit variance, if this is what you want. This is the default setting and actually doesn't need to expressed explicitly. Setting paramter ``cvType=["loo"]`` we make sure that we compute the PCR model using full cross validation. ``"loo"`` means "Leave One Out". By setting paramter ``numpComp=4`` we ask for four components to be computed.
+
[…]
