+++ examples/PCA_on_descriptive_sensory_analysis_data.py
@@ -1,0 +1,394 @@
+
+# coding: utf-8
+
+# # Principal component analysis (PCA) on descriptive sensory analysis data
+
+# This notebook illustrates how to use the **hoggorm** package to carry out principal component analysis (PCA) on a multivariate descriptive sensory analysis data from cheese. Furthermore, we will learn how to visualise the results of the PCA using the **hoggormPlot** package.
+
+# ---
+
+# ### Import packages and prepare data
+
+# First import **hoggorm** for analysis of the data and **hoggormPlot** for plotting of the analysis results. We'll also import **pandas** such that we can read the data into a data frame. **numpy** is needed for checking dimensions of the data.
+
+# In[1]:
+
+
+import hoggorm as ho
+import hoggormplot as hop
+import pandas as pd
+import numpy as np
+
+
+# Next, load the cancer data that we are going to analyse using **hoggorm**. After the data has been loaded into the pandas data frame, we'll display it in the notebook.
+
+# In[2]:
+
+
+# Load data
+
+# Insert code for reading data from other folder in repository instead of directly from same repository.
+data_df = pd.read_csv('cheese_sensory.txt', sep='\t', index_col=0)
+data_df
+
+
+# Let's have a look at the dimensions of the data frame.
+
+# In[3]:
+
+
+np.shape(data_df)
+
+
+# There are observations for 34 countries as well as all OECD countries together, which results in 35 rows. Furthermore, there are 10 columns where each column represents one type of cancer in men.
+
+# The ``nipalsPCA`` class in hoggorm accepts only **numpy** arrays with numerical values and not pandas data frames. Therefore, the pandas data frame holding the imported data needs to be "taken apart" into three parts: 
+# * a numpy array holding the numeric values
+# * a Python list holding variable (column) names
+# * a Python list holding object (row) names. 
+# 
+# The array with values will be used as input for the ``nipalsPCA`` class for analysis. The Python lists holding the variable and row names will be used later in the plotting function from the **hoggormPlot** package when visualising the results of the analysis. Below is the code needed to access both data, variable names and object names.
+
+# In[4]:
+
+
+# Get the values from the data frame
+data = data_df.values
+
+# Get the variable or columns names
+data_varNames = list(data_df.columns)
+
+# Get the object or row names
+data_objNames = list(data_df.index)
+
+
+# Let's have a quick look at the column or variable names.
+
+# In[5]:
+
+
+data_varNames
+
+
+# Now show the object or row names.
+
+# In[6]:
+
+
+data_objNames
+
+
+# ---
+
+# ### Apply PCA to our data
+
[…]
