### 2019-02-17 097e5d6
Added example for SMI, updated examples for PCR and PLSR
