### 2019-03-16 #980 @labarba
I see that no progress has been made on this submission since the reviewer comments from Nov 10, 2018. At this point, I will add the "paused" label.

@olivertomic — Please ping `@openjournals/joss-eics` when you're have made your enhancements and would like us to take a new look.
