### 2019-03-18 #980 @olivertomic
Dear @labarba - Thank you for your notice. We have made some progress, like adding some examples in Jupyter Notebooks (as requested by @jsgalan ), but admittedly the process has been painfully slow due to various reasons. We will ping `@openjournals/joss-eics` once we are done. Thank you again.
