### 2019-05-21 #980 @danielskatz
In my opinion, yes, if this is a firm commitment, but I will also defer to @arfon as the editor of the submission.
