### 2019-05-21 #980 @arfon
> Dear @danielskatz . Yes, we are still planning on making these changes. I am finally done teaching this semester and plan to finish this up by end of June. I hope this time frame is still OK.

👍 this sounds OK to me.
