### 2019-06-17 b46fa54
Update quickstart.rst
Fixed typo
