### 2019-06-17 849d349
Update README.rst
Fixed typo.
