### 2019-06-17 82b4a4f
Update README.rst
Fix typo.
