### 2019-06-18 a21d5ee
Data to right folder
Place all data files into the folder of each method.
