### 2019-06-18 fb8d4dd
Extended PCA examples
Added example illustrating use of PCA on spectroscopy data.
