+++ hoggorm/pca.py
@@ -123,15 +123,16 @@ class nipalsPCA:
         # size of CV segments)
         numObj = np.shape(self.arrX_input)[0]
         
+        
         # Compute the sizes of training sets in CV
         if self.cvType[0] == "loo":
-            print("loo")
+            print("Selected cross validation type: loo")
             cvComb = cv.LeaveOneOut(numObj)
         elif self.cvType[0] == "KFold":
-            print("KFold")
+            print("Selected cross validation type: KFold")
             cvComb = cv.KFold(numObj, k=self.cvType[1])
         elif self.cvType[0] == "lolo":
-            print("lolo")
+            print("Selected cross validation type: lolo")
             cvComb = cv.LeaveOneLabelOut(self.cvType[1])
         else:
             print('Requested form of cross validation is not available')
@@ -401,13 +402,13 @@ class nipalsPCA:
             numObj = np.shape(self.arrX)[0]
 
             if self.cvType[0] == "loo":
-                print("loo")
+                print("Selected cross validation type: loo")
                 cvComb = cv.LeaveOneOut(numObj)
             elif self.cvType[0] == "KFold":
-                print("KFold")
+                print("Selected cross validation type: KFold")
                 cvComb = cv.KFold(numObj, k=self.cvType[1])
             elif self.cvType[0] == "lolo":
-                print("lolo")
+                print("Selected cross validation type: lolo")
                 cvComb = cv.LeaveOneLabelOut(self.cvType[1])
             else:
                 print('Requested form of cross validation is not available')
