+++ examples/PCA/PCA_on_spectroscopy_data.py
@@ -1,0 +1,379 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+# # Principal component analysis (PCA) on spectroscopy data
+
+# This notebook illustrates how to use the **hoggorm** package to carry out principal component analysis (PCA) on spectroscopy data. Furthermore, we will learn how to visualise the results of the PCA using the **hoggormPlot** package.
+
+# ---
+
+# ### Import packages and prepare data
+
+# First import **hoggorm** for analysis of the data and **hoggormPlot** for plotting of the analysis results. We'll also import **pandas** such that we can read the data into a data frame. **numpy** is needed for checking dimensions of the data.
+
+# In[1]:
+
+
+import hoggorm as ho
+import hoggormplot as hop
+import pandas as pd
+import numpy as np
+
+
+# Next, load the spectroscopy data that we are going to analyse using **hoggorm**. After the data has been loaded into the pandas data frame, we'll display it in the notebook.
+
+# In[6]:
+
+
+# Load data
+
+# Insert code for reading data from other folder in repository instead of directly from same repository.
+data_df = pd.read_csv('gasoline_NIR.txt', header=None, sep='\s+')
+
+
+# Let's have a look at the dimensions of the data frame.
+
+# In[7]:
+
+
+np.shape(data_df)
+
+
+# The ``nipalsPCA`` class in hoggorm accepts only **numpy** arrays with numerical values and not pandas data frames. Therefore, the pandas data frame holding the imported data needs to be "taken apart" into three parts: 
+# * a numpy array holding the numeric values
+# * a Python list holding variable (column) names
+# * a Python list holding object (row) names. 
+# 
+# The array with values will be used as input for the ``nipalsPCA`` class for analysis. The Python lists holding the variable and row names will be used later in the plotting function from the **hoggormPlot** package when visualising the results of the analysis. Below is the code needed to access both data, variable names and object names.
+
+# In[10]:
+
+
+# Get the values from the data frame
+data = data_df.values
+
+
+# ---
+
+# ### Apply PCA to our data
+
+# Now, let's run PCA on the data using the ``nipalsPCA`` class. The documentation provides a [description of the input parameters](https://hoggorm.readthedocs.io/en/latest/pca.html). Using input paramter ``arrX`` we define which numpy array we would like to analyse. By setting input parameter ``Xstand=False`` we make sure that the variables are only mean centered, not scaled to unit variance. This is the default setting and actually doesn't need to expressed explicitly. Setting paramter ``cvType=["loo"]`` we make sure that we compute the PCA model using full cross validation. ``"loo"`` means "Leave One Out". By setting paramter ``numpComp=4`` we ask for four principal components (PC) to be computed.
+
+# In[12]:
+
+
+model = ho.nipalsPCA(arrX=data, Xstand=False, cvType=["loo"], numComp=5)
+
+
[…]

[…]
