### 2019-06-19 d55101f
Added PLSR example
Added PLSR example using gasoline data with NIR and octane measurements.
