+++ examples/PLSR/PLSR_on_NIR_and_octane_data.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-# # Partial Least Squares Regression (PLSR) on Sensory and Fluorescence data
+# # Partial Least Squares Regression (PLSR) on Near Infrared Spectroscopy (NIR) data and octane data
 
 # This notebook illustrates how to use the **hoggorm** package to carry out partial least squares regression (PLSR) on multivariate data. Furthermore, we will learn how to visualise the results of the PLSR using the **hoggormPlot** package.
 
@@ -11,7 +11,7 @@
 
 # First import **hoggorm** for analysis of the data and **hoggormPlot** for plotting of the analysis results. We'll also import **pandas** such that we can read the data into a data frame. **numpy** is needed for checking dimensions of the data.
 
-# In[2]:
+# In[1]:
 
 
 import hoggorm as ho
@@ -26,16 +26,16 @@ import numpy as np
 
 
 # Load fluorescence data
-X_df = pd.read_csv('cheese_fluorescence.txt', index_col=0, sep='\t')
+X_df = pd.read_csv('gasoline_NIR.txt', header=None, sep='\s+')
 X_df
 
 
-# In[4]:
+# In[6]:
 
 
-# Load sensory data
-Y_df = pd.read_csv('cheese_sensory.txt', index_col=0, sep='\t')
-Y_df
+# Load response data, that is octane measurements
+y_df = pd.read_csv('gasoline_octane.txt', header=None, sep='\s+')
+y_df
 
 
 # The ``nipalsPLS2`` class in hoggorm accepts only **numpy** arrays with numerical values and not pandas data frames. Therefore, the pandas data frames holding the imported data need to be "taken apart" into three parts: 
@@ -45,44 +45,44 @@ Y_df
 # 
 # The numpy arrays with values will be used as input for the ``nipalsPLS2`` class for analysis. The Python lists holding the variable and row names will be used later in the plotting function from the **hoggormPlot** package when visualising the results of the analysis. Below is the code needed to access both data, variable names and object names.
 
-# In[5]:
+# In[7]:
 
 
 # Get the values from the data frame
 X = X_df.values
-Y = Y_df.values
+y = y_df.values
 
 # Get the variable or columns names
 X_varNames = list(X_df.columns)
-Y_varNames = list(Y_df.columns)
+y_varNames = list(y_df.columns)
 
 # Get the object or row names
 X_objNames = list(X_df.index)
-Y_objNames = list(Y_df.index)
+y_objNames = list(y_df.index)
 
 
 # ---
 
 # ### Apply PLSR to our data
 
[…]
