+++ paper.bib
@@ -17,6 +17,17 @@
   isbn = {9780444863010}
 }
 
+@INBOOK{wold75,
+  chapter = {Path models with latent variables: The NIPALS approach},
+  title = {Quantitative Sociology: International perspectives on mathematical and statistical model building},
+  publisher = {Academic Press, NY},
+  year = {1975},
+  editor = {H.M. Blalock et al.},
+  author = {H. Wold},
+  pages = {307-357},
+  isbn = {0121039501}
+}
+
 @ARTICLE{naes88,
   author = {Tormod N{\ae}s, and Harald Martens},
   title = {Principal components regression in {NIR} analysis: Viewpoints, background details and selection of components},
