### 2019-06-22 d0a7653
Fixed license text
