### 2019-06-22 f1419f8
Updated readthedocs docu
Inserted new links for juypter notebook examples.
