+++ docs/quickstart.rst
@@ -7,7 +7,7 @@ hoggorm is a Python package for explorative multivariate statistics in Python. I
 * PCR (principal component regression)
 * PLSR (partial least squares regression)
   
-  - PLSR1 for single variable responses
+  - PLSR1 for univariate responses
   - PLSR2 for multivariate responses
 * matrix correlation coefficients RV and RV2.
 
@@ -65,14 +65,31 @@ Documentation
 - Jupyter notebooks with examples of how to use hoggorm
   
   - for `PCA`_
-  - for PCR (coming soon)
-  - for PLSR1 (coming soon)
-  - for PLSR2 (coming soon)
-  - for matrix correlation ceoefficitents RV and RV2 (coming soon)
+    - `PCA on cancer data`_ on men in OECD countries
+	- `PCA on NIR spectroscopy data`_ measured on gasoline	
+	- `PCA on sensory data`_ measured on cheese
+  - for `PCR`_
+    - PCR on NIR spectroscopy and octane data measured on gasoline (coming soon)
+	- `PCR on sensory and fluorescence spectroscopy data`_ measured on cheese
+  - for `PLSR1`_ for univariate response (one response variable)
+    - `PLSR1 on NIR spectroscopy and octane data`_ measured on gasoline
+  - for `PLSR2`_ for multivariate response (multiple response variables)
+    - `PLSR2 on sensory and fluorescence spectroscopy`_ measured on cheese
+  - for matrix correlation ceoefficitents `RV and RV2`_
   
 
 .. _Read the Docs: http://hoggorm.readthedocs.io/en/latest
-.. _PCA: https://github.com/olivertomic/hoggorm/blob/master/docs/PCA%20with%20hoggorm.ipynb
+.. _PCA: https://github.com/olivertomic/hoggorm/tree/master/examples/PCA
+.. _PCR: https://github.com/olivertomic/hoggorm/tree/master/examples/PCR
+.. _PLSR: https://github.com/olivertomic/hoggorm/tree/master/examples/PLSR
+.. _RV and RV2: https://github.com/olivertomic/hoggorm/tree/master/examples/RV_%26_RV2
+.. _PCA on cancer data: https://github.com/olivertomic/hoggorm/blob/master/examples/PCA/PCA_on_cancer_data.ipynb
+.. _PCA on NIR spectroscopy data: https://github.com/olivertomic/hoggorm/blob/master/examples/PCA/PCA_on_spectroscopy_data.ipynb
+.. _PCA on sensory data: https://github.com/olivertomic/hoggorm/blob/master/examples/PCA/PCA_on_descriptive_sensory_analysis_data.ipynb
+.. _PCR on sensory and fluorescence spectroscopy data: https://github.com/olivertomic/hoggorm/blob/master/examples/PCR/PCR_on_sensory_and_fluorescence_data.ipynb
+.. _PLSR1 on NIR spectroscopy and octane data: https://github.com/olivertomic/hoggorm/blob/master/examples/PLSR/PLSR_on_NIR_and_octane_data.ipynb
+.. _PLSR2 on sensory and fluorescence spectroscopy: https://github.com/olivertomic/hoggorm/blob/master/examples/PLSR/PLSR_on_sensory_and_fluorescence_data.ipynb
+.. _RV and RV2: https://github.com/olivertomic/hoggorm/blob/master/examples/RV_%26_RV2/RV_and_RV2_on_sensory_and_fluorescence_data.ipynb
 
 More examples in Jupyter notebooks are provided at `hoggormExamples GitHub repository`_.
 
