### 2019-06-22 2f62503
Update quickstart.rst
