### 2019-06-27 #980 @arfon
> @labarba @arfon @danielskatz I have solved most of the issues suggested by @jsgalan as can be seen from the repository updates and new Jupyter notebooks with examples. I would to ask you kindly for another week to finish up some polishing of the documentation. Would that be OK for you?

Yes, that's fine with me. Thanks for keeping with this!
