### 2019-06-28 c4ced0b
Update SMI example
