+++ examples/SMI/SMI example.py
@@ -1,4 +1,4 @@
-
+#!/usr/bin/env python
 # coding: utf-8
 
 # # SMI - Similarity of Matrices Index
@@ -14,15 +14,15 @@
 # `X1` is a matrix of size (100 x 300) generated by drawing 300000 pseudo-random numbers before column centring.  
 # `X2` is a copy of matrix `X1` where the 3rd principal component has been omitted, by performing singular value decomposition on `X1` and constructing `X2` without the 3rd left and right singular vectors and without the 3rd singular value.
 
-# In[6]:
+# In[1]:
 
 
 import numpy as np
 import hoggorm as ho
     
-X1 = ho.center(np.random.rand(100,300))
+X1 = ho.center(np.random.rand(100, 300))
 U, s, V = np.linalg.svd(X1, 0)
-X2 = np.dot(np.dot(np.delete(U, 2,1), np.diag(np.delete(s,2))), np.delete(V,2,0))
+X2 = np.dot(np.dot(np.delete(U, 2, 1), np.diag(np.delete(s, 2))), np.delete(V, 2, 0))
 X1.shape
 
 
@@ -32,8 +32,8 @@ X1.shape
 # In[2]:
 
 
-smiOP = ho.SMI(X1, X2, ncomp1 = 10, ncomp2 = 10)
-print(np.round(smiOP.smi,2))
+smiOP = ho.SMI(X1, X2, ncomp1=10, ncomp2=10)
+print(np.round(smiOP.smi, 2))
 
 
 # A hypothesis can be made regarding the similarity of two subspaces where the null hypothesis is that they are equal and the alternative is that they are not. Permutation testing yields the following P-values (probabilities that the observed difference could be larger given the null hypothesis is true).
@@ -41,31 +41,36 @@ print(np.round(smiOP.smi,2))
 # In[3]:
 
 
-print(np.round(smiOP.significance(),2))
+print(np.round(smiOP.significance(), 2))
 
 
-# In[12]:
+# In[4]:
 
 
-#import hoggormplot as hoPlot
-#hoPlot.plotSMI(smiOP,[10,10])
-import hoggormplot
-plotSMI(smiOP,[10,10])
+import hoggormplot as hop
+
+
+# In[5]:
+
+
+# import hoggormplot for plotting
+#import hoggormplot as hop
+hop.plotSMI(smiOP, [10, 10])
 
 
 # ### Procrustes Rotations
 # The similarities using PR <= OP, and in this simple case OP$^2$ = PR. Otherwise the pattern stays the same.
 
-# In[4]:
+# In[6]:
 
 
 smiPR = ho.SMI(X1, X2, ncomp1 = 10, ncomp2 = 10, projection = "Procrustes")
-print(np.round(smiPR.smi,2))
+print(np.round(smiPR.smi, 2))
 
 
 # The number of permutations can be controlled for quick (100) or accurate (>10000) computations of significance.
 
-# In[5]:
+# In[7]:
 
 
 print(np.round(smiPR.significance(B = 100)))
@@ -74,10 +79,10 @@ print(np.round(smiPR.significance(B = 100)))
 # ### Custom subspace
 # Instead of the PCA subspaces automatically used, one can supply a custom subspace, e.g. scores from Partial Least Squares, in the following way (here just copying the original scores from above):
 
-# In[6]:
+# In[8]:
 
 
-smiCustom = ho.SMI(X1, X2, ncomp1 = 10, ncomp2 = 10, Scores1 = U)
+smiCustom = ho.SMI(X1, X2, ncomp1=10, ncomp2=10, Scores1=U)
 
 
 # _Reference:_   
