### 2019-06-28 5dd683b
Update docu SMI
