### 2019-06-28 06a88f2
Update SMI example
