+++ examples/SMI/SMI_pseudo-random_numbers.py
@@ -3,12 +3,12 @@
 
 # # SMI - Similarity of Matrices Index
 # SMI is a measure of the similarity between the dominant subspaces of two matrices. It comes in two flavours (projections): 
-# - OP - Orthogonal Projections, and 
+# - OP - Orthogonal Projections
 # - PR - Procrustes Rotations.  
 # 
 # The former (default) compares subspaces using ordinary least squares and can be formulated as the explained variance  when predicting one matrix subspace using the other matrix subspace. PR is a restriction where only rotation and scaling is allowed in the similarity calculations.  
 #   
-# Subspaces are by default computed using Principal Component Analysis (PCA). When the number of components extracted from one of the matrices is smaller than the other, the explained variance is calculate predicting the smaller subspace using the larger subspace.
+# Subspaces are by default computed using Principal Component Analysis (PCA). When the number of components extracted from one of the matrices is smaller than the other, the explained variance is calculated predicting the smaller subspace by using the larger subspace.
 
 # ## Example: Structured differences between random matrices
 # `X1` is a matrix of size (100 x 300) generated by drawing 300000 pseudo-random numbers before column centring.  
@@ -29,6 +29,8 @@ X1.shape
 # ### Orthogonal Projections
 # Removing the 3rd principal component from one of the matrices, one would expect that the first two components are exactly equal. The third component should yield an explained variance equal to the proportion explained, i.e. 2/3, while the following components will have an increasing proportion of explained variance (SMI). Rows correspond to the first matrix and collumns correspond to the second matrix in the prints below. An extra subspace dimension in `X1` will always cover the extra component in `X2`, while the lost component in `X2` can never be compensated for, thus the pattern below.
 
+# More details regarding the use of the SMI are found in the [documentation](https://hoggorm.readthedocs.io/en/latest/matrix_corr_coeff.html).
+
 # In[2]:
 
 
@@ -44,27 +46,29 @@ print(np.round(smiOP.smi, 2))
 print(np.round(smiOP.significance(), 2))
 
 
+# Now import hoggormplot package and plot similarities.
+
 # In[4]:
 
 
+# import hoggormplot for plotting
 import hoggormplot as hop
 
 
 # In[5]:
 
 
-# import hoggormplot for plotting
-#import hoggormplot as hop
+# Plot similarities
 hop.plotSMI(smiOP, [10, 10])
 
 
 # ### Procrustes Rotations
 # The similarities using PR <= OP, and in this simple case OP$^2$ = PR. Otherwise the pattern stays the same.
 
-# In[6]:
+# In[9]:
 
 
-smiPR = ho.SMI(X1, X2, ncomp1 = 10, ncomp2 = 10, projection = "Procrustes")
+smiPR = ho.SMI(X1, X2, ncomp1=10, ncomp2=10, projection="Procrustes")
 print(np.round(smiPR.smi, 2))
 
 
