### 2019-06-28 21dcf31
Update SMI example
