### 2019-07-02 6d5d2c9
Update SMI example
