+++ examples/SMI/SMI_pseudo-random_numbers.py
@@ -14,7 +14,7 @@
 # `X1` is a matrix of size (100 x 300) generated by drawing 300000 pseudo-random numbers before column centring.  
 # `X2` is a copy of matrix `X1` where the 3rd principal component has been omitted, by performing singular value decomposition on `X1` and constructing `X2` without the 3rd left and right singular vectors and without the 3rd singular value.
 
-# In[1]:
+# In[2]:
 
 
 import numpy as np
@@ -31,7 +31,7 @@ X1.shape
 
 # More details regarding the use of the SMI are found in the [documentation](https://hoggorm.readthedocs.io/en/latest/matrix_corr_coeff.html).
 
-# In[2]:
+# In[3]:
 
 
 smiOP = ho.SMI(X1, X2, ncomp1=10, ncomp2=10)
@@ -40,7 +40,7 @@ print(np.round(smiOP.smi, 2))
 
 # A hypothesis can be made regarding the similarity of two subspaces where the null hypothesis is that they are equal and the alternative is that they are not. Permutation testing yields the following P-values (probabilities that the observed difference could be larger given the null hypothesis is true).
 
-# In[3]:
+# In[4]:
 
 
 print(np.round(smiOP.significance(), 2))
@@ -48,24 +48,26 @@ print(np.round(smiOP.significance(), 2))
 
 # Now import hoggormplot package and plot similarities.
 
-# In[4]:
+# In[5]:
 
 
 # import hoggormplot for plotting
 import hoggormplot as hop
 
 
-# In[5]:
+# In[6]:
 
 
 # Plot similarities
 hop.plotSMI(smiOP, [10, 10])
 
 
+# The significance symbols in the diamond plot above indicate if a chosen subspace from one matrix can be found inside the subspace from the other matrix ($\supset$, $\subset$, =), or if there is signficant difference (P-values <0.001\*\*\* <0.01 \*\* <0.05 \* <0.1 . >=0.1).
+
 # ### Procrustes Rotations
 # The similarities using PR <= OP, and in this simple case OP$^2$ = PR. Otherwise the pattern stays the same.
 
-# In[9]:
+# In[7]:
 
 
 smiPR = ho.SMI(X1, X2, ncomp1=10, ncomp2=10, projection="Procrustes")
@@ -74,10 +76,10 @@ print(np.round(smiPR.smi, 2))
 
 # The number of permutations can be controlled for quick (100) or accurate (>10000) computations of significance.
 
-# In[7]:
+# In[8]:
 
 
-print(np.round(smiPR.significance(B = 100)))
+print(np.round(smiPR.significance(B = 100),2))
 
 
 # ### Custom subspace
