+++ tests/test_pls1.py
@@ -1,25 +1,10 @@
 # -*- coding: utf-8 -*-
 """
-Created on Fri Jun 29 13:45:30 2018
-
-@author: olive
+Test whether PLS1 results are as expected.
 """
-
-'''
-FIXME: PCA testing ideas:
- * Well known datasets (iris)
- * Combinations of input parameters
- * Edge case datasets
- * Big matrix for performance testing / profiling
- * Illegale data and error handling (zero variance)
- * Integer and float type matrix
-'''
 import os.path as osp
-
 import numpy as np
-
 import pytest
-
 from hoggorm import nipalsPLS1 as PLS1
 
 
@@ -57,7 +42,7 @@ ATTRS = [
     'X_MSECV',
     'X_RMSECV_indVar',
     'X_RMSECV',
-    #'X_scores_predict',
+    'X_scores_predict',
     'Y_means',
     'Y_loadings',
     'Y_corrLoadings',
@@ -79,24 +64,91 @@ ATTRS = [
 ]
 
 
+@pytest.fixture(scope="module")
+def pls1cached(cfldat, csecol2dat):
+    """
+    Run PLS1 from current hoggorm installation and compare results against reference results.
+    """
+    return PLS1(arrX=cfldat, vecy=csecol2dat, cvType=["loo"])
+
+
+testMethods = ["X_scores", "X_loadings", "X_corrLoadings", "X_cumCalExplVar_indVar", "X_cumCalExplVar",
+               "Y_loadings", "Y_corrLoadings", "Y_cumCalExplVar"] 
+# testMethods = ["Y_corrLoadings"]
+@pytest.fixture(params=testMethods)
+def pls1ref(request, datafolder):
+    """
+    Load reference numerical results from file.
+    """
+    rname = request.param
+    refn = "ref_PLS1_{}.tsv".format(rname.lower())
+    try:
+        refdat = np.loadtxt(osp.join(datafolder, refn))
+    except FileNotFoundError:
+        refdat = None
+
+    return (rname, refdat)
+
+
+def test_compare_reference(pls1ref, pls1cached):
+    """
+    Check whether numerical outputs are the same (or close enough).
+    """
+    rname, refdat = pls1ref
+    res = getattr(pls1cached, rname)()
+
+    if refdat is None:
+        dump_res(rname, res)
+        assert False, "Missing reference data for {}, data is dumped".format(rname)
+    elif rname == 'Y_corrLoadings' or rname == 'Y_loadings':
+        if not np.allclose(res[:3], refdat.reshape(1, -1)[:3], rtol=rtol, atol=atol):
+            dump_res(rname, res)
+            assert False, "Difference in {}, data is dumped".format(rname)    
+    elif rname == 'X_cumCalExplVar' or rname == 'Y_cumCalExplVar':
+        if not np.allclose(np.array(res[:3]), refdat[:3], rtol=rtol, atol=atol):
+            dump_res(rname, res)
+            assert False, "Difference in {}, data is dumped".format(rname)
+    elif not np.allclose(res[:, :3], refdat[:, :3], rtol=rtol, atol=atol):
+        dump_res(rname, res)
+        assert False, "Difference in {}, data is dumped".format(rname)
+    else:
+        assert True
+
+
+def dump_res(rname, dat):
+    """
+    Dumps information to file if reference data is missing or difference is larger than tolerance.
+    """
+    dumpfolder = osp.realpath(osp.dirname(__file__))
+    dumpfn = "dump_PLS1_{}.tsv".format(rname.lower())
+    np.savetxt(osp.join(dumpfolder, dumpfn), dat, fmt='%.9e', delimiter='\t')
+
+
 def test_api_verify(pls1cached, cfldat):
     """
[…]
