### 2019-07-04 7411a63
Updated text and references
