+++ paper.bib
@@ -7,25 +7,16 @@
   isbn = {9780124712508}
 }
 
-@INBOOK{wold82,
-  chapter = {Soft modelling: The basics and some extensions},
-  title = {Systems under Indirect Observation},
-  publisher = {Amsterdam: North Holland},
-  year = {1982},
-  editor = {K.G. J{\"o}reskog and H. Wold},
-  author = {H. Wold},
-  isbn = {9780444863010}
-}
-
-@INBOOK{wold75,
-  chapter = {Path models with latent variables: The NIPALS approach},
-  title = {Quantitative Sociology: International perspectives on mathematical and statistical model building},
-  publisher = {Academic Press, NY},
-  year = {1975},
-  editor = {H.M. Blalock et al.},
-  author = {H. Wold},
-  pages = {307-357},
-  isbn = {0121039501}
+@INBOOK{wold83,
+  title = {Matrix Pencils. Lecture Notes in Mathematics},
+  chapter = {The multivariate calibration problem in chemistry solved by the PLS method},
+  pages = {286-293},
+  publisher = {Springer, Berlin, Heidelberg},
+  year = {1983},
+  editor = {B. K\aagstr\om and A. Ruhe},
+  author = {H. Wold, M. Martens, S. Wold},
+  isbn = {9783540119838}
+  doi = {https://doi.org/10.1007/BFb0062108}
 }
 
 @ARTICLE{naes88,
@@ -47,7 +38,6 @@
   pages = {257–265}, 
   doi = {10.2307/2347233}
 }
-
 
 @ARTICLE{smilde09,
   author = {A.K. Smilde and H.A.L. Kiers  and S. Bijlsma and C.M. Rubingh and M.J. van Erk},
