+++ paper.bib
@@ -15,7 +15,7 @@
   year = {1983},
   editor = {B. K\aagstr\om and A. Ruhe},
   author = {H. Wold, M. Martens, S. Wold},
-  isbn = {9783540119838}
+  isbn = {9783540119838},
   doi = {https://doi.org/10.1007/BFb0062108}
 }
 
