+++ paper.bib
@@ -13,7 +13,7 @@
   pages = {286-293},
   publisher = {Springer, Berlin, Heidelberg},
   year = {1983},
-  editor = {B. K\aagstr\om and A. Ruhe},
+  editor = {B. K\aa gstr\o m and A. Ruhe},
   author = {H. Wold, M. Martens, S. Wold},
   isbn = {9783540119838},
   doi = {https://doi.org/10.1007/BFb0062108}
