### 2019-07-04 b847242
Update quickstart.rst
