### 2019-07-04 52b31b0
Update quickstart.rst
