### 2019-07-06 ee4e7f2
Updated examples
