+++ examples/PLSR/PLSR_on_sensory_and_fluorescence_data.py
@@ -11,7 +11,7 @@
 
 # First import **hoggorm** for analysis of the data and **hoggormPlot** for plotting of the analysis results. We'll also import **pandas** such that we can read the data into a data frame. **numpy** is needed for checking dimensions of the data.
 
-# In[2]:
+# In[1]:
 
 
 import hoggorm as ho
@@ -22,7 +22,7 @@ import numpy as np
 
 # Next, load the data that we are going to analyse using **hoggorm**. After the data has been loaded into the pandas data frame, we'll display it in the notebook.
 
-# In[3]:
+# In[2]:
 
 
 # Load fluorescence data
@@ -30,7 +30,7 @@ X_df = pd.read_csv('cheese_fluorescence.txt', index_col=0, sep='\t')
 X_df
 
 
-# In[4]:
+# In[3]:
 
 
 # Load sensory data
@@ -45,7 +45,7 @@ Y_df
 # 
 # The numpy arrays with values will be used as input for the ``nipalsPLS2`` class for analysis. The Python lists holding the variable and row names will be used later in the plotting function from the **hoggormPlot** package when visualising the results of the analysis. Below is the code needed to access both data, variable names and object names.
 
-# In[5]:
+# In[4]:
 
 
 # Get the values from the data frame
@@ -67,7 +67,7 @@ Y_objNames = list(Y_df.index)
 
 # Now, let's run PLSR on the data using the ``nipalsPLS2`` class. The documentation provides a [description of the input parameters](https://hoggorm.readthedocs.io/en/latest/plsr.html). Using input paramter ``arrX`` and ``arrY`` we define which numpy array we would like to analyse. ``arrY`` is what typically is considered to be the response matrix, while the measurements are typically defined as ``arrX``. By setting input parameter ``Xstand=False`` and ``Ystand=False`` we make sure that the variables are only mean centered, not scaled to unit variance, if this is what you want. This is the default setting and actually doesn't need to expressed explicitly. Setting paramter ``cvType=["loo"]`` we make sure that we compute the PLS2 model using full cross validation. ``"loo"`` means "Leave One Out". By setting paramter ``numpComp=4`` we ask for four components to be computed.
 
-# In[6]:
+# In[5]:
 
 
 model = ho.nipalsPLS2(arrX=X, Xstand=False, 
@@ -78,7 +78,7 @@ model = ho.nipalsPLS2(arrX=X, Xstand=False,
 
[…]
