+++ examples/PCR/PCR_on_sensory_and_fluorescence_data.py
@@ -1,4 +1,4 @@
-
+#!/usr/bin/env python
 # coding: utf-8
 
 # # Principal Component Regression (PCR) on Sensory and Fluorescence data
@@ -45,7 +45,7 @@ Y_df
 # 
 # The numpy arrays with values will be used as input for the ``nipalsPCR`` class for analysis. The Python lists holding the variable and row names will be used later in the plotting function from the **hoggormPlot** package when visualising the results of the analysis. Below is the code needed to access both data, variable names and object names.
 
-# In[5]:
+# In[4]:
 
 
 # Get the values from the data frame
@@ -67,7 +67,7 @@ Y_objNames = list(Y_df.index)
 
 # Now, let's run PCR on the data using the ``nipalsPCR`` class. The documentation provides a [description of the input parameters](https://hoggorm.readthedocs.io/en/latest/pcr.html). Using input paramter ``arrX`` and ``arrY`` we define which numpy array we would like to analyse. ``arrY`` is what typically is considered to be the response matrix, while the measurements are typically defined as ``arrX``. By setting input parameter ``Xstand=False`` and ``Ystand=False`` we make sure that the variables are only mean centered, not scaled to unit variance, if this is what you want. This is the default setting and actually doesn't need to expressed explicitly. Setting paramter ``cvType=["loo"]`` we make sure that we compute the PCR model using full cross validation. ``"loo"`` means "Leave One Out". By setting paramter ``numpComp=4`` we ask for four components to be computed.
 
-# In[6]:
+# In[5]:
 
 
 model = ho.pcr.nipalsPCR(arrX=X, Xstand=False, 
@@ -78,7 +78,7 @@ model = ho.pcr.nipalsPCR(arrX=X, Xstand=False,
 
 # That's it, the PCR model has been computed. Now we would like to inspect the results by visualising them. We can do this using plotting functions of the separate [**hoggormPlot** package](https://hoggormplot.readthedocs.io/en/latest/). If we wish to plot the results for component 1 and component 2, we can do this by setting the input argument ``comp=[1, 2]``. The input argument ``plots=[1, 2, 3, 4, 6]`` lets the user define which plots are to be plotted. If this list for example contains value ``1``, the function will generate the scores plot for the model. If the list contains value ``2``, then the loadings plot will be plotted. Value ``3`` stands for correlation loadings plot and value ``4`` stands for bi-plot and ``6`` stands for explained variance plot. The hoggormPlot documentation provides a [description of input paramters](https://hoggormplot.readthedocs.io/en/latest/mainPlot.html).
 
-# In[7]:
+# In[6]:
 
 
 hop.plot(model, comp=[1, 2], 
@@ -90,51 +90,51 @@ hop.plot(model, comp=[1, 2],
 
 # Plots can also be called separately.
 
-# In[8]:
+# In[7]:
 
 
[…]
