+++ examples/PLSR/PLSR_on_NIR_and_octane_data.py
@@ -22,7 +22,7 @@ import numpy as np
 
 # Next, load the data that we are going to analyse using **hoggorm**. After the data has been loaded into the pandas data frame, we'll display it in the notebook.
 
-# In[3]:
+# In[2]:
 
 
 # Load fluorescence data
@@ -30,7 +30,7 @@ X_df = pd.read_csv('gasoline_NIR.txt', header=None, sep='\s+')
 X_df
 
 
-# In[6]:
+# In[3]:
 
 
 # Load response data, that is octane measurements
@@ -38,14 +38,14 @@ y_df = pd.read_csv('gasoline_octane.txt', header=None, sep='\s+')
 y_df
 
 
-# The ``nipalsPLS2`` class in hoggorm accepts only **numpy** arrays with numerical values and not pandas data frames. Therefore, the pandas data frames holding the imported data need to be "taken apart" into three parts: 
+# The ``nipalsPLS1`` class in hoggorm accepts only **numpy** arrays with numerical values and not pandas data frames. Therefore, the pandas data frames holding the imported data need to be "taken apart" into three parts: 
 # * two numpy array holding the numeric values
 # * two Python list holding variable (column) names
 # * two Python list holding object (row) names. 
 # 
 # The numpy arrays with values will be used as input for the ``nipalsPLS2`` class for analysis. The Python lists holding the variable and row names will be used later in the plotting function from the **hoggormPlot** package when visualising the results of the analysis. Below is the code needed to access both data, variable names and object names.
 
-# In[7]:
+# In[4]:
 
 
 # Get the values from the data frame
@@ -67,61 +67,61 @@ y_objNames = list(y_df.index)
 
 # Now, let's run PLSR on the data using the ``nipalsPLS1`` class, since we have a univariate response. The documentation provides a [description of the input parameters](https://hoggorm.readthedocs.io/en/latest/plsr.html). Using input paramter ``arrX`` and ``vecy`` we define which numpy array we would like to analyse. ``vecy`` is what typically is considered to be the response vector, while the measurements are typically defined as ``arrX``. By setting input parameter ``Xstand=False`` we make sure that the variables are only mean centered, not scaled to unit variance, if this is what you want. This is the default setting and actually doesn't need to expressed explicitly. Setting paramter ``cvType=["loo"]`` we make sure that we compute the PLS2 model using full cross validation. ``"loo"`` means "Leave One Out". By setting paramter ``numpComp=10`` we ask for four components to be computed.
 
-# In[9]:
+# In[5]:
 
 
 model = ho.nipalsPLS1(arrX=X, Xstand=False, 
-                      vecy=Y,
+                      vecy=y,
                       cvType=["loo"], 
                       numComp=10)
 
 
[…]
