+++ paper.md
@@ -36,7 +36,7 @@ Currently (version 0.13.0), statistical methods implemented in hoggorm are: (I) 
 
 The hoggorm package provides access to an extended repertiore of interpretation tools that are integrated in PCA, PCR, PLS1 and PLS2. These including scores, loadings, correlation loadings, explained variances for calibrated and validated models (both for individual variables as well as all variables together). Scores are the objects' coordinates in the compressed data representation and can for instance be used to search for patterns or groups among the objects. Loadings are the variables' representations in the compressed space showing their contribution to the components. Finally, correlation loadings show how each variable correlates to the score vectors/components and how much of the variation in each variable is explained across components. Note that models trained with hoggorm may also be applied for prediction purposes, both for continuous and categorical variables, where appropriate.
 
-Furthermore, hoggorm implements the matrix correlation coefficient methods RV [@robert76] and RV2 (also known as modified RV) [@smilde09], as well as the similarity index for comparing coupled matrices index (SMI) [@indahl18]. These methods can be used to quickly determine how much common information there is between two data matrices.
+Furthermore, hoggorm implements the matrix correlation coefficient methods RV [@robert76] and RV2 (also known as modified RV) [@smilde09], as well as the similarity index for comparing coupled matrices index (SMI) [@indahl18]. These methods can be used to quickly determine how much common information there is between two data matrices. Results from models trained with hoggorm may be visualised using the complementary plotting package hoggormplot
 
 # Acknowledgements
 Both users and developers have made valuable contributions to improve the usability the hoggorm library. This includes reporting of bugs, testing various features and other forms of feedback. A complete list of contributors is provided at https://github.com/olivertomic/hoggorm/graphs/contributors
