### 2019-07-06 7efc1de
Update paper.md
