### 2019-07-06 884717a
Update paper.bib
