+++ paper.bib
@@ -14,7 +14,7 @@
   publisher = {Springer, Berlin, Heidelberg},
   year = {1983},
   editor = {B. K\aa gstr\"{o}m and A. Ruhe},
-  author = {H. Wold, M. Martens, S. Wold},
+  author = {H. Wold and M. Martens and S. Wold},
   isbn = {9783540119838},
   doi = {https://doi.org/10.1007/BFb0062108}
 }
