### 2019-07-06 ff76c85
Updates for paper
