### 2019-07-06 #980 @olivertomic
Dear @jsgalan and @arfon . I think we have addressed now all issues raised. We have done the following:

1. Added Jupyter notebooks and corresponding python scripts (exports from the Jupyter notebooks) that provide working examples of how to use PCA, PCR, PLS, SMI, RV and RV2.

2. Where appropriate, we implemented examples on different type of data. The Jupyter notebooks for each method are stored in their respective folders. All those folders are stored in the example folder.

3. In those Jupter notebooks we added more examples of how to make single plots and how to generate plots that contain multiple subplots.

4. Documentation was updated according to changes.

5. Information on required packages for hoggorm and hoggormplot were improved. We tested hoggorm and hoggorm plot on several computers and could not reproduce issues with closing plot windows.

6. The test scripts were extended and explanation was added to the documentation of hogggorm (in the quickstart page)

7. We fixed the license.

8. We extended the article to describe better what the implemented methods can be used for and what tools the user can use for interpretation of the models.

9. The reference was fixed.

10. All these changes (and some bugfixes in the complementary hoggormplot package) were implemented and collected in new packages hoggorm (v0.13.2) and hoggormplot (v0.13.2). The packages were uploaded to pypi.org, such that users can pip install them.

Thank you again for your patience. Please let us know whether there are any other issues that we can take care of.

best
Oliver
