### 2019-07-09 #980 @jsgalan
Hi all

I just did a fresh install of the package and got it installed perfectly

Also tried some of the examples without no issues

The license was  completed

Documentation hosted in https://hoggorm.readthedocs.io/en/latest/quickstart.html looks good and the examples Jupyter notebooks are well explained.

I will just suggest a small change in the example in the documentation, since the example is not complete (**as is**), it will be helpful to have a working version.

Also the article is complete with the appropriate references.

@arfon  my part I think is ready to be published.

Best luck to all
