### 2019-07-10 10b95f6
Update quickstart.rst
