### 2019-07-10 cb804c9
Update quickstart.rst
