+++ docs/quickstart.rst
@@ -125,27 +125,27 @@ Example
 	>>> scores = model.X_scores()
 	>>> print(scores)
 	[[-0.97535198 -1.71827581  0.43672952]
- 	 [ 1.28340424 -0.24453505 -0.98250731]
- 	 [-0.9127492   0.97132275  1.04708189]
- 	 [ 2.34954599  0.30633998  0.43178679]
- 	 [-1.74484905  0.68514813 -0.93309089]]
+	 [ 1.28340424 -0.24453505 -0.98250731]
+	 [-0.9127492   0.97132275  1.04708189]
+	 [ 2.34954599  0.30633998  0.43178679]
+	 [-1.74484905  0.68514813 -0.93309089]]
 	
 	# Get PCA loadings
 	>>> loadings = model.X_loadings()
 	>>> print(loadings)
 	[[ 0.55080115  0.10025801  0.25045298]
-     [ 0.57184198 -0.11712858  0.00316316]
-     [ 0.57141459  0.00568809  0.10503941]
-     [-0.1682551  -0.61149788  0.77153937]
-     [ 0.12161589 -0.77605877 -0.57528864]]
+	 [ 0.57184198 -0.11712858  0.00316316]
+	 [ 0.57141459  0.00568809  0.10503941]
+	 [-0.1682551  -0.61149788  0.77153937]
+	 [ 0.12161589 -0.77605877 -0.57528864]]
 	
 	# Get cumulative explained variance for each variable
 	>>> cumCalExplVar_allVariables = model.X_cumCalExplVar_indVar()
 	>>> print(cumCalExplVar_allVariables)
 	[[ 0.          0.          0.          0.          0.        ]
-     [90.98654597 98.07234952 97.92497156  8.48956314  4.43690992]
-     [92.12195756 99.62227118 97.92862256 50.73769558 72.47502242]
-     [97.31181824 99.62309922 98.84150821 99.98958248 99.85786661]]
+	 [90.98654597 98.07234952 97.92497156  8.48956314  4.43690992]
+	 [92.12195756 99.62227118 97.92862256 50.73769558 72.47502242]
+	 [97.31181824 99.62309922 98.84150821 99.98958248 99.85786661]]
 	
 	# Get cumulative explained variance for all variables
 	>>> cumCalExplVar_total = model.X_cumValExplVar()
