### 2019-07-10 5241cc9
Update quickstart.rst
