+++ docs/quickstart.rst
@@ -26,7 +26,7 @@ Make sure that Python 3.5 or higher is installed. A convenient way to install Py
 
 .. _Anaconda distribution: https://www.anaconda.com/download/
 
-- numpy
+- numpy >= 1.11.3
 
 
 Installation and upgrades
