### 2019-07-10 7d6361b
Update quickstart.rst
