### 2019-07-10 ed92db3
Update paper.md
