+++ paper.md
@@ -39,6 +39,6 @@ The hoggorm package provides access to an extended repertoire of interpretation 
 Furthermore, hoggorm implements the matrix correlation coefficient methods RV [@robert76] and RV2 (also known as modified RV) [@smilde09], as well as the similarity index for comparing coupled matrices index (SMI) [@indahl18]. These methods can be used to quickly determine how much common information there is between two data matrices. Results from models trained with hoggorm may be visualised using the complementary plotting package hoggormplot (v0.13.2).
 
 # Acknowledgements
-Both users and developers have made valuable contributions to improve the usability the hoggorm library. This includes reporting of bugs, testing various features and other forms of feedback. A complete list of contributors is provided at https://github.com/olivertomic/hoggorm/graphs/contributors. The authors are also thankful to reviewer Dr. Javier Sanchez Galan at the Universidad Tecnológica de Panama in Panama for constructive feedback on requirements for publication the hoggorm package at the Journal of Open Source Software.
+Both users and developers have made valuable contributions to improve the usability the hoggorm library. This includes reporting of bugs, testing various features and other forms of feedback. A complete list of contributors is provided at https://github.com/olivertomic/hoggorm/graphs/contributors. The authors are also thankful to reviewer Dr. Javier Sanchez Galan at the Universidad Tecnológica de Panama in Panama for testing and constructive feedback on requirements for publication of the hoggorm package at the Journal of Open Source Software.
 
 # References
