### 2019-07-11 #980 @olivertomic
Thanks a lot to both of you, @jsgalan (for taking on the review and testing job) and @arfon (for guidance and patience). I have made the requested changes in the hoggorm package and made an archive of hoggorm at Zenodo.

10.5281/zenodo.3326328
http://doi.org/10.5281/zenodo.3326328

Looking very much forward to get the paper published!
